Tweakable: Add Tweakable component
This component is designed to tie in with the props passed for
each variable in InterpolatedAnimation, where we have a visualization
of a value that is "tweakable" in the sense that it can be edited
by dragging the mouse cursor. The editor will be appended to the
end of the children of the element.